Move everything to internal/utils & move ComputeVtxoTaprootScript to common
